rename: shared-game types 파일 이름 수정
