+++ app/explore/page.tsx
@@ -1,7 +1,7 @@
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
 import { getSharedGames } from '@/entities/shared-game/model/use-shared-games.query';
-import type { SortKey } from '@/entities/shared-game/model/types';
+import type { SortKey } from '@/entities/shared-game/model/shared-game.type';
 
 import Header from '@/widgets/header/ui/header';
 import TagFilter from '@/features/tag-filter/ui/tag-filter';
